+++ celt/pvq_search_asm.go
@@ -3,8 +3,7 @@
 package celt
 
 // pvqSearchBestPos finds the position with the best rate-distortion score
-// for placing a pulse. xy and yy are passed as float64 to avoid float32
-// stack layout issues; they are converted to float32 inside the assembly.
+// for placing a pulse.
 //
 // The function computes:
 //
@@ -12,7 +11,7 @@ package celt
 //	for j = 1..n-1: if bestDen*(xy+absX[j])^2 > (yy+y[j])*bestNum: update best
 //
 //go:noescape
-func pvqSearchBestPos(absX, y []float32, xy, yy float64, n int) int
+func pvqSearchBestPos(absX, y []float32, xy, yy float32, n int) int
 
 // pvqSearchPulseLoop places pulsesLeft pulses one at a time into the best
 // position using the rate-distortion criterion. It merges the entire outer
@@ -24,14 +23,14 @@ func pvqSearchBestPos(absX, y []float32, xy, yy float64, n int) int
 //	absX[0..n-1] = absolute values of input vector (read-only)
 //	y[0..n-1]    = 2*iy[j] pulse counts (modified in-place: y[bestID] += 2 per pulse)
 //	iy[0..n-1]   = integer pulse counts (modified in-place: iy[bestID]++ per pulse)
-//	xy, yy       = running cross-correlation and energy (float32 via float64 args)
+//	xy, yy       = running cross-correlation and energy
 //	n            = vector dimension
 //	pulsesLeft   = number of pulses to place
 //
-// Returns updated (xy, yy) as float64 (containing float32 values).
+// Returns updated (xy, yy).
 //
 //go:noescape
-func pvqSearchPulseLoop(absX, y []float32, iy []int, xy, yy float64, n, pulsesLeft int) (float64, float64)
+func pvqSearchPulseLoop(absX, y []float32, iy []int, xy, yy float32, n, pulsesLeft int) (float32, float32)
 
 //go:noescape
 func pvqExtractAbsSignAsm(x []float64, absX []float32, y []float32, signx []byte, iy []int, n int)
